requirements.txt
python-telegram-bot==20.3
apscheduler==3.10.10
fastapi==0.111.0
uvicorn==0.23.2
